Basic class extend change
