+++ src/EVE/BattleTask.py
@@ -1,0 +1,6 @@
+import EVETask
+
+class BattleTask(EVETask):
+  def __init__(self, hwnd, index, eveRunners):
+    super().__init__(hwnd, index, eveRunners)
+
