+++ src/EVE/mainThread.py
@@ -5,7 +5,7 @@ import os
 sys.path.append(os.path.abspath(__file__ + "\\..\\..\\utils"))
 sys.path.append(os.path.abspath(__file__ + "\\..\\"))
 from windows import *
-from EVETask import EVETask
+from MiningTask import MiningTask
 import datetime as dt
 
 # todo
@@ -14,17 +14,15 @@ import datetime as dt
 # Keep time difference by 60s all the time. Done partial
 # 1280x720
 
-
 def isWorkHour():
     hour = dt.datetime.now().hour
     if hour >= 5 and hour <= 9:
         return False
     return True
 
-
 def runTask(hwnd, index):
-    #mode 0 standard, mode 1 for weekly high safety mining
-    task = EVETask(hwnd, index, "MuMuPlayer",mode=0)
+    #mode 0 standard, mode 1 for weekly high safety mining, mode 2 for gulan target mining
+    task = MiningTask(hwnd, index,mode=0)
     while True:
         try:
             if not (isWorkHour()):
@@ -37,7 +35,6 @@ def runTask(hwnd, index):
             task.closeWindow()
             print("thread failed, stop")
             raise (e)
-
 
 def main():
     print("开工前todo list: 打开本地列表,v船到广角,驻地最后一个,改变战斗列表为名称排序")
@@ -57,7 +54,6 @@ def main():
         )
     for thread in threads:
         thread.start()
-
 
 if __name__ == "__main__":
     main()
